refactor(common): share the bounded-ingest primitive, keep limit vocabularies local
ADR-043 §1 applies to two verifiers that live in different crates. The evidence
bundle reader has a ceiling mechanism; `assay_core::replay` has none at all. Both
need the same primitive, and duplicating it would leave two implementations of a
boundary that is already easy to get wrong.

`LimitReader` therefore moves to `assay-common::limits`, behind the `std` feature
because it is an `io::Read` adapter and the crate is `no_std` for the eBPF side.
`assay-evidence` re-exports it so existing call sites are unchanged.

Only the mechanism moves. `VerifyLimits` stays in `assay-evidence`: its eight
fields describe an evidence bundle, `max_manifest_bytes` and `max_events` say
nothing about a replay bundle, and that vocabulary does not belong in a shared
`no_std` crate. Replay will get its own small limit set over the same reader.
One mechanism, several vocabularies.

This adds an `assay-evidence -> assay-common` edge, which did not exist before.
`assay-common` is a leaf with no internal dependencies, so there is no cycle, and
the edge is far lighter than the alternative of pulling `assay-evidence` and its
tar, flate2, ed25519 and uuid dependencies into `assay-core` for one struct. The
documented crate graph in CLAUDE.md is updated in the same commit, with a note on
what the edge does and does not carry.

Verified: `assay-common` still builds with `--no-default-features`, clippy is
clean with `-D warnings` across both crates, and the full `assay-evidence` suite
passes unchanged. No behaviour change; the boundary correction it builds on is
the preceding commit.
